fix(toolbox): normalize guided progress value type

+++ lib/features/guided_journey/widgets/guided_journey_common_widgets.dart
@@ -57,7 +57,7 @@ class GuidedJourneyProgressStrip extends StatelessWidget {
           ClipRRect(
             borderRadius: BorderRadius.circular(999),
             child: LinearProgressIndicator(
-              value: progress.clamp(0, 1),
+              value: progress.clamp(0.0, 1.0).toDouble(),
               minHeight: 9,
               backgroundColor: const Color(0xFFE5E7EB),
               valueColor:
